[libcompat/W32U] minizip_iowin32u.c: Use W32U_alloca.h functions to convert strings.
Note that extra casting is required because the open64 prototype
uses const void*, whereas we need const char*. We can't simply
cast it in the macro parameter because that screws up the macro.

Marked the internal functions as static.

Removed the notes about UTF-8, since W32U_IsUnicode() handles
the system UTF-8 check for us.