[gens-qt4] translations/*.ts: Updated the translation files using lupdate.sh.
Strings related to UnRAR and external programs in general have been
removed due to the libarchive changes.

The only new strings added are system abbreviations. I'm not sure if they
actually need to be localized, or if they do, I should probably add a
separate unlocalized version. These abbreviations are displayed in the
title bar to indicate the system that is currently being emulated.