[libgens] Disabled the Decompressor subsystem.
The RAR and 7zip decompressors need to be ported to Archive.
For now, I'll just disable the Decompressor subsystem.

Note that this breaks gens-qt4, since it stores the RAR program/DLL
in the DcRar class, and #including any Decompressor headers results
in an error due to an #error I added to Decompressor.hpp.

Archive.hpp, Rom.hpp: Minor changes to adjust for this change.