[unrar] Initial commit of UnRAR 5.3.9.
On Windows systems, UnRAR won't be built at all; however, we will
use dll.hpp for the interface. Instead, the Gens/GS II zip file
will contain an original signed copy of UnRAR.dll (or UnRAR64.dll).

On other systems, I will build a copy of UnRAR as a shared library,
libgensunrar.so. It will have the same interface as UnRAR.dll on
Windows (though using the C calling convention instead of PASCAL).
It will be dlopen()'d at runtime to avoid having to directly
link to it, which should help with licensing issues.