[unrar] dll.hpp: #define _UNIX on Unix, Linux, and Mac OS X.
This is usually defined by raros.hpp when building UnRAR.dll,
but it isn't accessible by users of UnRAR.dll.

We could just define _UNIX globally on the top level, but I haven't
seen this macro used anywhere else, so #define _UNIX here if it
isn't defined and the standard Unix, Linux, or Mac OS X macros
are present.