[libgensfile] Rar: New function openRom().
This function first checks the RAR magic (which works for RARv2
and later), and then attempts to open the file using UnRAR.dll.

FIXME: UnRAR.dll uses mbsrtowcs(), which is dependent on the system
locale. It also doesn't work properly for me; it fails at the first
non-ASCII character. I'll need to convert the strings to UTF-16
using libgenstext and then use the UTF-16 functions.