Call setlocale(LC_ALL, "") on program startup.
This updates gens-qt4, gens-sdl, and mcd_pcm.

This is needed in order for UnRAR.dll to handle UTF-8 properly on
Linux and Unix, since the wide character functions aren't initialized
if we don't set the locale.

An interesting side effect is this causes gens-sdl's OSD to show
timestamps with 4-digit years and 12-hour times on en_US, whereas
previously it was showing 2-digit years and 24-hour times.

I tested en_GB as well, and it shows DD/MM/YY format.

NOTE: Remember to use UTF-8 locales (e.g. en_US.UTF-8); otherwise,
UnRAR.dll won't be able to handle Unicode filenames properly.

NOTE 2: Test suites have NOT been updated to use setlocale().
This is intentional, and it helps to ensure that tests don't
randomly break due to a locale change.