[libgensfile] Rar::readFile(): Implemented file reading.
I can now load a ROM with a non-ASCII filename from a RAR file
with a non-ASCII filename on Linux. The RAR was also created
on Linux.

On Windows, I can load a ROM with an ASCII filename from a RAR file
with a non-ASCII filename. If the ROM has a non-ASCII filename,
the load fails due to ANSI issues.