[libgensfile] Rar::getFileInfo(): Fixed Unicode filename handling on Windows.
UnRAR.dll always ensures the Unicode field is filled out,
so we don't have to worry about checking the ANSI field.

Similarly, on Linux, UnRAR.dll ensures that the Unicode name
is converted to UTF-8 (or whatever the system locale is set to),
so we can just strdup() that field directly.

This fixes loading a ROM file with a Unicode name from a RAR archive.