[libgensfile] LzmaSdk: New base class for LZMA SDK-based archive handlers.
This base class handles some initialization requird by all archive
handlers that use the LZMA SDK, including CRC tables and filename
encoding conversion.

NOTE: The Crc64 table is initialized regardless of whether or not
Xz is used. It's currently only used by Xz right now, but it could
be used by 7-Zip later as well. (Or, it *is* used by 7-Zip for LZMA2
streams, but I'm not sure.)

Better to just initialize all of the CRC tables regardless so we
don't encounter weird errors later.