[libgensfile] Lzma: Preliminary Lzma archive class.
Currently, it only supports detecting the Lzma archive magic, and only
if the archive was created using xz-utils. The file format doesn't
have a really good magic number compared to .xz or .7z.

Note that .lzma wasn't widely used, so I don't expect the autodetection
screwiness to be an issue.

TODO: Add an option to disable autodetection of certain formats to
allow the user to work around a false positive if one occurs.